pyproject.toml
[project]
name = "geometrydashclone"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pillow==11.2.1",
    "pygame==2.6.1",
]
